moving from /user to /me
